python-requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
cryptography==3.1
decorator==4.4.2
Django==3.1
django-admin-interface==0.13.4
django-colorfield==0.3.2
django-elasticsearch-dsl==7.1.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
elasticsearch==7.9.1
elasticsearch-dsl==7.2.1
IMDbPY==2020.9.29
install==1.3.3
lxml==4.5.2
pycparser==2.20
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
